Created new component to encapsulate schema, semantic type, and unit controls.
