+++ src/app/property/property.component.html
@@ -18,38 +18,7 @@
     <display-name-description [formIndex]="formIndex[1]" [parentFormGroup]="property.form"></display-name-description>
     <br/>
 
-    <mat-form-field color="accent">
-      <mat-label attr.for="property_schema_{{formIndex}}">Schema</mat-label>
-      <mat-select [compareWith]="compareSchemas" matNativeControl id="property_schema_{{formIndex}}" 
-                  (selectionChange)="changeSchema($event)" [formControl]="schemaDropDownControl" #schema>
-        <mat-option *ngFor="let schemaType of getSchemaTypes().filter(validSchemaTypes)" [value]="schemaType">{{schemaType}}</mat-option>
-      </mat-select>
-    </mat-form-field>
-    <span class="formSpacer"></span>
-
-    <ng-container *ngIf="isComplex(schema.value)">
-      <button mat-raised-button color="primary" (click)="openSchemaEditor()">
-          <mat-icon>edit</mat-icon>
-          Edit Schema
-      </button>
-    </ng-container>
-    <br/>
-
-    <mat-form-field color="accent">
-        <mat-label attr.for="property_semanticType_{{formIndex}}">Semantic Type</mat-label>
-        <mat-select matNativeControl id="property_semanticType_{{formIndex}}"
-                    (selectionChange)="changeSemanticType($event)" [formControl]="semanticTypeDropDownControl" #semanticType>
-            <mat-option *ngFor="let semanticType of getSemanticTypes()" [value]="semanticType">{{semanticType}}</mat-option>
-        </mat-select>
-    </mat-form-field>
-    <span class="formSpacer"></span>
-
-    <mat-form-field color="accent">
-        <mat-label attr.for="property_unit_{{formIndex}}">Unit</mat-label>
-        <mat-select matNativeControl id="property_unit_{{formIndex}}" formControlName="unit">
-            <mat-option *ngFor="let unit of getUnits()" [value]="unit"> {{ unit }} </mat-option>
-        </mat-select>
-    </mat-form-field>   
+    <schema-semantic-type-unit type="property" [formIndex]="formIndex[1]" [parentForm]="property"></schema-semantic-type-unit>
     <br/>
 
     <mat-form-field color="accent" style="width: 630px">
