+++ src/app/property/property.component.ts
@@ -1,11 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { EditorService } from '../services/editor/editor.service';
-import { MatSelectChange } from '@angular/material/select';
 import { MatDialog } from '@angular/material/dialog';
-import { SchemaService } from '../services/schema/schema.service';
 import { PropertyCapabilityFormControl } from '../formControls/PropertyCapabilityFormControl';
 import { AbstractCapabilityFormControl } from '../formControls/AbstractCapabilityFormControl';
-import { SchemaTypeEnum } from '../models/SchemaTypeEnum';
 import { UntypedFormControl } from '@angular/forms';
 import { AbstractSchemaModel } from '../models/AbstractSchemaModel';
 
@@ -18,16 +14,12 @@ export class PropertyComponent implements OnInit {
   @Input() public formIndex!: [number, number];
   @Input() public property!: PropertyCapabilityFormControl;
   @Input() public panelOpenState!: boolean;
-  private _editorService: EditorService;
-  private _schemaService: SchemaService;
   public schemaFormControl!: AbstractCapabilityFormControl<AbstractSchemaModel> | undefined;
   public schemaDropDownControl: UntypedFormControl = new UntypedFormControl();
   public semanticTypeDropDownControl: UntypedFormControl = new UntypedFormControl();
   public dialog: MatDialog;
 
-  constructor(editorService: EditorService, schemaService: SchemaService, dialog: MatDialog) {
-    this._editorService = editorService;
-    this._schemaService = schemaService;
+  constructor(dialog: MatDialog) {
     this.dialog = dialog;
   }
 
@@ -63,97 +55,5 @@ export class PropertyComponent implements OnInit {
     name?.valueChanges.subscribe(value => {
       name.setValue(value, { emitEvent: false })
     });
-  }
-
-  public getSchemaTypes(): Array<string> {
-    return this._schemaService.getSchemaTypes();
-  }
-
-  public getSemanticTypes(): Array<string> {
-    return this._editorService.getSemanticTypes();
-  }
-
-  public getUnits(): string[] | undefined {
-    let semanticType = this.semanticTypeDropDownControl?.value;
-    if(semanticType) {
-      let units = this._editorService.getUnits().get(semanticType);
-      return units;
-    }
-
-    return undefined;
-  }
-
-  public getFormControl(name: string): UntypedFormControl {
-    return this.property.form.get(name) as UntypedFormControl;
-  }
-
-  // TODO: Passing validSchemaTypes to the FilterPipe doesn't get re-evaluated on changes
-  //       Ideally, we'd use the FilterPipe in the component view to filter the array values.
-  //       However, the filter function only executes once and is not ever re-evaluated if the view or form changes. 
-  public validSchemaTypes = (schemaType: string): boolean => {
-    if (this.semanticTypeDropDownControl?.value)
-      return ["double", "float", "integer", "long"].indexOf(schemaType.toLowerCase()) > -1;
-    else
-      return true;
-  }
-
-  public changeSemanticType($event: MatSelectChange): void {
-    let value = $event.value;
-    this.changeSemanticTypeInternal(value);
-  }
-
-  private changeSemanticTypeInternal(value: string): void {
-    let type = this.property.form.get("type");
-
-    if (["", null, undefined].indexOf(value) > -1) {
-      let semanticType = new Array<string>("Property");
-      type?.setValue(semanticType);
-      let unit = this.property.form.get("unit");
-      unit?.setValue(undefined);
-    } else {
-      let semanticType = new Array<string>("Property", value);
-      type?.setValue(semanticType);
-
-      let schema = this.property.form.get("schema")?.value;
-      if (["double", "float", "integer", "long"].indexOf(schema.toLowerCase()) === -1) {
-        this.property.form.get("schema")?.setValue(undefined);
-        this.schemaDropDownControl.setValue(undefined);
-        this.schemaFormControl = undefined;
-      }
-    }
-  }
-
-  public isComplex(schema: string): boolean {
-    if (typeof schema == 'string')
-      return this._schemaService.getSchemaTypeEnum(schema) == SchemaTypeEnum.Complex;
-
-    return false;
-  }
-
-  public compareSchemas = (model1: AbstractSchemaModel, model2: AbstractSchemaModel): boolean => {
-    return this._schemaService.compareSchemas(model1, model2)
-  }
-
-  public changeSchema($event: MatSelectChange): void {
-    let value = $event.value;
-    this.changeSchemaInternal(value);
-  }
-
-  public changeSchemaInternal(value: any): void {
-    if (value instanceof AbstractCapabilityFormControl) return;
-    let key = value.toLowerCase();
-    let schemaType = this._schemaService.getSchemaTypeEnum(key);
-    this.property.form.get("schema")?.setValue(key);
-
-    if (schemaType == SchemaTypeEnum.Complex) {
-      let formControl = this._schemaService.createForm(SchemaTypeEnum[schemaType], key);
-      if (formControl === undefined) return;
-      this.schemaFormControl = formControl;
-    }
-  }
-
-  public openSchemaEditor(): void {
-    if(this.schemaFormControl)
-      this._schemaService.openSchemaEditor(this.property, this.schemaFormControl)
   }
 }
