+++ src/app/schema-semantic-type-unit/schema-semantic-type-unit.component.html
@@ -1,0 +1,35 @@
+<ng-container [formGroup]="parentForm.form">
+  <mat-form-field color="accent">
+    <mat-label attr.for="{{type}}_schema_{{formIndex}}">Schema</mat-label>
+    <mat-select [compareWith]="compareSchemas" matNativeControl id="{{type}}_schema_{{formIndex}}"
+      (selectionChange)="changeSchema($event)" [formControl]="schemaDropDownControl" #schema>
+      <mat-option *ngFor="let schemaType of getSchemaTypes().filter(validSchemaTypes)"
+        [value]="schemaType">{{schemaType}}</mat-option>
+    </mat-select>
+  </mat-form-field>
+  <span class="formSpacer"></span>
+
+  <ng-container *ngIf="isComplex(schema.value)">
+    <button mat-raised-button color="primary" (click)="openSchemaEditor()">
+      <mat-icon>edit</mat-icon>
+      Edit Schema
+    </button>
+  </ng-container>
+  <br />
+
+  <mat-form-field color="accent">
+    <mat-label attr.for="{{type}}_semanticType_{{formIndex}}">Semantic Type</mat-label>
+    <mat-select matNativeControl id="{{type}}_semanticType_{{formIndex}}" (selectionChange)="changeSemanticType($event)"
+      [formControl]="semanticTypeDropDownControl" #semanticType>
+      <mat-option *ngFor="let semanticType of getSemanticTypes()" [value]="semanticType">{{semanticType}}</mat-option>
+    </mat-select>
+  </mat-form-field>
+  <span class="formSpacer"></span>
+
+  <mat-form-field color="accent">
+    <mat-label attr.for="{{type}}_unit_{{formIndex}}">Unit</mat-label>
+    <mat-select matNativeControl id="{{type}}_unit_{{formIndex}}" formControlName="unit">
+      <mat-option *ngFor="let unit of getUnits()" [value]="unit"> {{ unit }} </mat-option>
+    </mat-select>
+  </mat-form-field>
+</ng-container>
