+++ src/app/telemetry/telemetry.component.ts
@@ -1,12 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { EditorService } from '../services/editor/editor.service';
-import { MatSelectChange } from '@angular/material/select';
 import { MatDialog } from '@angular/material/dialog';
-import { SchemaService } from '../services/schema/schema.service';
 import { TelemetryCapabilityFormControl } from '../formControls/TelemetryCapabilityFormControl';
 import { AbstractCapabilityFormControl } from '../formControls/AbstractCapabilityFormControl';
-import { AbstractCapabilityModel } from '../models/AbstractCapabilityModel';
-import { SchemaTypeEnum } from '../models/SchemaTypeEnum';
 import { UntypedFormControl } from '@angular/forms';
 import { AbstractSchemaModel } from '../models/AbstractSchemaModel';
 
@@ -19,16 +14,12 @@ export class TelemetryComponent implements OnInit {
   @Input() public formIndex!: [number, number];
   @Input() public telemetry!: TelemetryCapabilityFormControl;
   @Input() public panelOpenState!: boolean;
-  private _editorService: EditorService;
-  private _schemaService: SchemaService;
   public schemaFormControl?: AbstractCapabilityFormControl<AbstractSchemaModel>;
   public schemaDropDownControl: UntypedFormControl = new UntypedFormControl();
   public semanticTypeDropDownControl: UntypedFormControl = new UntypedFormControl();
   public dialog: MatDialog;
 
-  constructor(editorService: EditorService, schemaService: SchemaService, dialog: MatDialog) {
-    this._editorService = editorService;
-    this._schemaService = schemaService;
+  constructor(dialog: MatDialog) {
     this.dialog = dialog;
   }
 
@@ -64,94 +55,5 @@ export class TelemetryComponent implements OnInit {
     name?.valueChanges.subscribe(value => {
       name.setValue(value, { emitEvent: false })
     });
-  }
-
-  public getSchemaTypes(): Array<string> {
-    return this._schemaService.getSchemaTypes();
-  }
-
-  public getSemanticTypes(): Array<string> {
-    return this._editorService.getSemanticTypes();
-  }
-
-  public getUnits(): string[] | undefined {
-    let semanticType = this.semanticTypeDropDownControl?.value;
-    if(semanticType) {
-      let units = this._editorService.getUnits().get(semanticType);
-      return units;
-    }
-
-    return undefined;
-  }
-
-  public getFormControl(name: string): UntypedFormControl {
-    return this.telemetry.form.get(name) as UntypedFormControl;
-  }
-
-  public validSchemaTypes = (schemaType: string): boolean => {
-    if (this.semanticTypeDropDownControl?.value)
-      return ["double", "float", "integer", "long"].indexOf(schemaType.toLowerCase()) > -1;
-    else
-      return true;
-  }
-
-  public changeSemanticType($event: MatSelectChange): void {
-    let value = $event.value;
-    this.changeSemanticTypeInternal(value);
-  }
-
-  private changeSemanticTypeInternal(value: string): void {
-    let type = this.telemetry.form.get("type");
-
-    if (["", null, undefined].indexOf(value) > -1) {
-      let semanticType = new Array<string>("Telemetry");
-      type?.setValue(semanticType);
-      let unit = this.telemetry.form.get("unit");
-      unit?.setValue(undefined);
-    } else {
-      let semanticType = new Array<string>("Telemetry", value);
-      type?.setValue(semanticType);
-
-      let schema = this.telemetry.form.get("schema")?.value;
-      if (["double", "float", "integer", "long"].indexOf(schema.toLowerCase()) === -1) {
-        this.telemetry.form.get("schema")?.setValue(undefined);
-        this.schemaDropDownControl.setValue(undefined);
-        this.schemaFormControl = undefined;
-      }
-    }
-  }
-
-  public isComplex(schema: string): boolean {
-    if (typeof schema == 'string')
-      return this._schemaService.getSchemaTypeEnum(schema) == SchemaTypeEnum.Complex;
-
-    return false;
-  }
-
-  public compareSchemas = (model1: AbstractSchemaModel, model2: AbstractSchemaModel): boolean => {
-    return this._schemaService.compareSchemas(model1, model2)
-  }
-
-  public changeSchema($event: MatSelectChange): void {
-    let value = $event.value;
-    this.changeSchemaInternal(value);
-  }
-
-  public changeSchemaInternal(value: any): void {
-    if (value instanceof AbstractCapabilityFormControl) return;
-    let key = value.toLowerCase();
-    let schemaType = this._schemaService.getSchemaTypeEnum(key);
-    this.telemetry.form.get("schema")?.setValue(key);
-
-    if (schemaType == SchemaTypeEnum.Complex) {
-      let formControl = this._schemaService.createForm(SchemaTypeEnum[schemaType], key);
-      if (formControl === undefined) return;
-      this.schemaFormControl = formControl;
-    }
-  }
-
-  public openSchemaEditor(): void {
-    if (this.schemaFormControl)
-      this._schemaService.openSchemaEditor(this.telemetry, this.schemaFormControl)
   }
 }
